Fix MSVC compiler warnings
reg.: conversion from 'size_t' to 'int', possible loss of data (C4267)

+++ src/importexport/guitarpro/internal/guitarbendimport/benddatacollector.cpp
@@ -251,7 +251,7 @@ void BendDataCollector::fillBendDataContext(BendDataContext& bendDataCtx)
                 for (size_t i = 0; i < notesVec.size(); ++i) {
                     Note* n = notesVec[i];
                     if (auto it = notesData.find(n); it != notesData.end()) {
-                        fillBendDataForNote(bendDataCtx, it->second, i);
+                        fillBendDataForNote(bendDataCtx, it->second, static_cast<int>(i));
                     }
                 }
             }
